combine dependency managers
